requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
pywa==3.8.0
gunicorn==21.2.0
